+++ NEWS.md
@@ -1,5 +1,13 @@
-# Version 0.6.1  
-- First Release
+# glitter 0.6.2.9000  
+
+* Add filter to `github api` functions for only the github user's repos  
+* Add function that retrieves tag/release history for a github user's repo  
+
+
+# glitter 0.6.1  
+
+* First Release  
+